Extract simple data table and refresh click
