+++ src/features/instance/config/roles/index.tsx
@@ -1,18 +1,17 @@
 import { Loading } from '@/components/Loading';
-import { Suspense, useCallback, useMemo, useState } from 'react';
+import { SimpleBrowseDataTable } from '@/components/SimpleBrowseDataTable';
+import { Button } from '@/components/ui/button';
+import { dataTableColumns } from '@/features/instance/config/roles/constants/tableDefinition';
+import { AddRoleModal } from '@/features/instance/config/roles/modals/AddRoleModal';
 import { getListRolesQueryOptions } from '@/features/instance/operations/queries/getListRoles';
+import { useRefreshClick } from '@/hooks/useRefreshClick';
+import { LocalRole } from '@/lib/api.patch';
 import { useSuspenseQuery } from '@tanstack/react-query';
 import { getRouteApi, useNavigate } from '@tanstack/react-router';
-import { BrowseDataTable } from '@/features/instance/config/roles/components/BrowseDataTable';
-import { dataTableColumns } from '@/features/instance/config/roles/constants/tableDefinition';
-import { EditRoleModal } from './modals/EditRoleModal';
-import { LocalRole } from '@/lib/api.patch';
 import { Row } from '@tanstack/react-table';
-import { Button } from '@/components/ui/button';
 import { PlusIcon, RefreshCwIcon } from 'lucide-react';
-import { toast } from 'sonner';
-import { sleep } from '@/lib/sleep';
-import { AddRoleModal } from '@/features/instance/config/roles/modals/AddRoleModal';
+import { Suspense, useCallback, useMemo, useState } from 'react';
+import { EditRoleModal } from './modals/EditRoleModal';
 
 const route = getRouteApi('');
 
@@ -59,20 +58,11 @@ export function ConfigRolesIndex() {
 		onSelectRole(undefined);
 	}, [onSelectRole]);
 
-	const onRefreshClick = useCallback(async () => {
-		const toastId = toast.loading('Refreshing...');
-		const startedAt = Date.now();
-		await refetch();
-		if (Date.now() - startedAt < 500) {
-			await sleep(500);
-		}
-		toast.dismiss(toastId);
-		toast.success('Refreshed!');
-	}, [refetch]);
+	const onRefreshClick = useRefreshClick(refetch);
 
 	return (
 		<Suspense fallback={<Loading className="flex flex-col items-center justify-center h-full" text="Loading..." />}>
-			<BrowseDataTable columns={dataTableColumns} data={localRoles} isFetching={isFetching} onRowClick={onRowClick}>
+			<SimpleBrowseDataTable columns={dataTableColumns} data={localRoles} isFetching={isFetching} onRowClick={onRowClick}>
 				<Button variant="defaultOutline" onClick={onRefreshClick} accessKey="r" disabled={isFetching || isRefetching}>
 					<RefreshCwIcon />
 					<span className="hidden lg:inline-block">
@@ -85,7 +75,7 @@ export function ConfigRolesIndex() {
 						<u>A</u>dd
 					</span>
 				</Button>
-			</BrowseDataTable>
+			</SimpleBrowseDataTable>
 			<AddRoleModal isModalOpen={isAddModalOpen} onChangesSaved={onRoleAdded} setIsModalOpen={setIsAddModalOpen} />
 			{isEditModalOpen && (
 				<EditRoleModal
